Introduce Interfaces for Test Data Handling and Observing
Added TestDataUploader and TestObserver interfaces to allow custom implementations and also improve flexibility and testability.
