+++ collector/unit-test-collector/src/main/kotlin/com/buildkite/test/collector/android/util/ConfigureTestUploader.kt
@@ -1,12 +1,13 @@
 package com.buildkite.test.collector.android.util
 
+import com.buildkite.test.collector.android.BuildKiteTestDataUploader
 import com.buildkite.test.collector.android.TestDataUploader
 import com.buildkite.test.collector.android.tracer.environment.BuildkiteEnvironmentValues
 
 /**
  * Configures a [TestDataUploader] instance for unit tests.
  */
-fun configureUnitTestUploader() = TestDataUploader(
+fun configureUnitTestUploader() = BuildKiteTestDataUploader(
     testSuiteApiToken = getStringEnvironmentValue(key = BuildkiteEnvironmentValues.BUILDKITE_ANALYTICS_TOKEN),
     isDebugEnabled = getBooleanEnvironmentValue(key = BuildkiteEnvironmentValues.BUILDKITE_ANALYTICS_DEBUG_ENABLED)
 )
